+++ openjdk8/linux/build.sh
@@ -38,7 +38,7 @@ function ensure_ant()
   if [ ! -x $OBF_DROP_DIR/ant/bin/ant ]; then
     mkdir -p $OBF_DROP_DIR/ant
     pushd $OBF_DROP_DIR/ant
-    curl -L http://mirrors.ircam.fr/pub/apache/ant/binaries/apache-ant-1.8.4-bin.tar.gz -o apache-ant-1.8.4-bin.tar.gz
+    curl -L http://archive.apache.org/dist/ant/binaries/apache-ant-1.8.4-bin.tar.gz -o apache-ant-1.8.4-bin.tar.gz
     tar xzf apache-ant-1.8.4-bin.tar.gz
     mv apache-ant-1.8.4/* .
     rmdir apache-ant-1.8.4
@@ -47,6 +47,7 @@ function ensure_ant()
   fi
 
   export PATH=$OBF_DROP_DIR/ant/bin:$PATH
+  export ANT_HOME=$OBF_DROP_DIR/ant
 }
 
 function ensure_cacert()
@@ -115,6 +116,14 @@ function ensure_java7()
       echo "missing required Java 7, aborting..."
     fi
     
+  elif [ "$CPU_BUILD_ARCH" = "ppc64" ]; then
+
+    if [ -d /opt/obuildfactory/jdk-1.7.0-openjdk-ppc64 ]; then
+      export OBF_BOOTDIR=/opt/obuildfactory/jdk-1.7.0-openjdk-ppc64
+    else
+      echo "missing required Java 7, aborting..."
+    fi
+
   else
 
     if [ -d /opt/obuildfactory/jdk-1.7.0-openjdk-i686 ]; then
@@ -134,6 +143,7 @@ function build_old()
   echo "### using old build system ###"
   
   NUM_CPUS=`grep "processor" /proc/cpuinfo | sort -u | wc -l`
+  [ $NUM_CPUS -gt 8 ] && NUM_CPUS=8
 
   export BUILD_NUMBER="$OBF_BUILD_DATE"
   export MILESTONE="$OBF_MILESTONE"
@@ -158,12 +168,14 @@ function build_old()
   
   if [ "$CPU_BUILD_ARCH" = "x86_64" ]; then
     export IMAGE_BUILD_DIR=$OBF_SOURCES_PATH/build/linux-amd64
+  elif [ "$CPU_BUILD_ARCH" = "ppc64" ]; then
+    export IMAGE_BUILD_DIR=$OBF_SOURCES_PATH/build/linux-ppc64
   else
     export IMAGE_BUILD_DIR=$OBF_SOURCES_PATH/build/linux-i586
   fi
 
   if [ "$XCLEAN" = "true" ]; then
-	  rm -rf $IMAGE_BUILD_DIR
+	rm -rf $IMAGE_BUILD_DIR
   fi
   
   # Set Company Name to OBuildFactory
@@ -185,8 +197,10 @@ function build_new()
   pushd $OBF_SOURCES_PATH/common/makefiles >>/dev/null
   
   # patch common/autoconf/version.numbers
-  mv ../autoconf/version.numbers ../autoconf/version.numbers.orig 
-  cat ../autoconf/version.numbers.orig | grep -v "MILESTONE" | grep -v "JDK_BUILD_NUMBER" | grep -v "COMPANY_NAME" > ../autoconf/version.numbers
+  if [ -f ../autoconf/version.numbers ]; then
+    mv ../autoconf/version.numbers ../autoconf/version.numbers.orig 
+    cat ../autoconf/version.numbers.orig | grep -v "MILESTONE" | grep -v "JDK_BUILD_NUMBER" | grep -v "COMPANY_NAME" > ../autoconf/version.numbers
+  fi
 
   export JDK_BUILD_NUMBER=$OBF_BUILD_DATE
   export MILESTONE=$OBF_MILESTONE
@@ -200,21 +214,29 @@ function build_new()
 
 	  if [ "$CPU_BUILD_ARCH" = "x86_64" ]; then
 	    BUILD_PROFILE=linux-x86_64-normal-server-fastdebug
+	  elif [ "$CPU_BUILD_ARCH" = "ppc64" ]; then
+	    BUILD_PROFILE=linux-ppc64-normal-server-fastdebug
+        EXTRA_FLAGS="--with-jvm-interpreter=cpp"
 	  else
   	    BUILD_PROFILE=linux-x86-normal-server-fastdebug
 	  fi
   
-	  sh ../autoconf/configure --with-boot-jdk=$OBF_BOOTDIR --with-freetype=$OBF_DROP_DIR/freetype --with-cacerts-file=$OBF_DROP_DIR/cacerts --with-ccache-dir=$OBF_WORKSPACE_PATH/.ccache --enable-debug
+	  bash ../autoconf/configure --with-boot-jdk=$OBF_BOOTDIR --with-freetype=$OBF_DROP_DIR/freetype --with-cacerts-file=$OBF_DROP_DIR/cacerts --with-ccache-dir=$OBF_WORKSPACE_PATH/.ccache --enable-debug \
+                               --with-build-number=$OBF_BUILD_DATE --with-milestone=$OBF_MILESTONE $EXTRA_FLAGS
 
   else
 
 	  if [ "$CPU_BUILD_ARCH" = "x86_64" ]; then
 	    BUILD_PROFILE=linux-x86_64-normal-server-release
+	  elif [ "$CPU_BUILD_ARCH" = "ppc64" ]; then
+	    BUILD_PROFILE=linux-ppc64-normal-server-release
+        EXTRA_FLAGS="--with-jvm-interpreter=cpp"
 	  else
 	    BUILD_PROFILE=linux-x86-normal-server-release
 	  fi
   
-	  sh ../autoconf/configure --with-boot-jdk=$OBF_BOOTDIR --with-freetype=$OBF_FREETYPE_DIR --with-cacerts-file=$OBF_DROP_DIR/cacerts --with-ccache-dir=$OBF_WORKSPACE_PATH/.ccache
+	  bash ../autoconf/configure --with-boot-jdk=$OBF_BOOTDIR --with-freetype=$OBF_FREETYPE_DIR --with-cacerts-file=$OBF_DROP_DIR/cacerts --with-ccache-dir=$OBF_WORKSPACE_PATH/.ccache \
+                               --with-build-number=$OBF_BUILD_DATE --with-milestone=$OBF_MILESTONE $EXTRA_FLAGS
 
   fi
 
@@ -227,7 +249,9 @@ function build_new()
   CONT=$BUILD_PROFILE make images
 
   # restore original common/autoconf/version.numbers
-  mv ../autoconf/version.numbers.orig ../autoconf/version.numbers
+  if [ -f ../autoconf/version.numbers.orig ]; then
+    mv ../autoconf/version.numbers.orig ../autoconf/version.numbers
+  fi
 
   popd >>/dev/null
 }
@@ -278,7 +302,7 @@ function archive_build()
 # Build start here
 #
 
-CPU_BUILD_ARCH=`uname -p`
+CPU_BUILD_ARCH=`uname -m`
 
 export JDK_BUNDLE_VENDOR="OBuildFactory"
 export BUNDLE_VENDOR="OBuildFactory"
